+++ src/ContaoManager/Plugin.php
@@ -1,7 +1,7 @@
 <?php
 
 /*
- * Copyright (c) 2020 Heimrich & Hannot GmbH
+ * Copyright (c) 2021 Heimrich & Hannot GmbH
  *
  * @license LGPL-3.0-or-later
  */
@@ -13,13 +13,10 @@ use Contao\ManagerPlugin\Bundle\BundlePluginInterface;
 use Contao\ManagerPlugin\Bundle\Config\BundleConfig;
 use Contao\ManagerPlugin\Bundle\Parser\ParserInterface;
 use Contao\ManagerPlugin\Config\ConfigPluginInterface;
-use Contao\ManagerPlugin\Config\ContainerBuilder;
-use Contao\ManagerPlugin\Config\ExtensionPluginInterface;
 use HeimrichHannot\MultiFileUploadBundle\HeimrichHannotContaoMultiFileUploadBundle;
-use HeimrichHannot\UtilsBundle\Container\ContainerUtil;
 use Symfony\Component\Config\Loader\LoaderInterface;
 
-class Plugin implements BundlePluginInterface, ExtensionPluginInterface, ConfigPluginInterface
+class Plugin implements BundlePluginInterface, ConfigPluginInterface
 {
     public function getBundles(ParserInterface $parser)
     {
@@ -37,31 +34,15 @@ class Plugin implements BundlePluginInterface, ExtensionPluginInterface, ConfigP
     }
 
     /**
-     * Allows a plugin to override extension configuration.
-     *
-     * @param string           $extensionName
-     * @param array            $extensionConfigs
-     * @param ContainerBuilder $container
-     *
-     * @return
-     */
-    public function getExtensionConfig($extensionName, array $extensionConfigs, ContainerBuilder $container)
-    {
-        return ContainerUtil::mergeConfigFile(
-            'huh_encore',
-            $extensionName,
-            $extensionConfigs,
-            __DIR__.'/../Resources/config/config_encore.yml'
-        );
-    }
-
-    /**
      * {@inheritdoc}
      */
     public function registerContainerConfiguration(LoaderInterface $loader, array $managerConfig)
     {
         $loader->load('@HeimrichHannotContaoMultiFileUploadBundle/Resources/config/services.yml');
-        $loader->load('@HeimrichHannotContaoMultiFileUploadBundle/Resources/config/hooks.yml');
         $loader->load('@HeimrichHannotContaoMultiFileUploadBundle/Resources/config/parameters.yml');
+
+        if (class_exists("HeimrichHannot\EncoreBundle\HeimrichHannotContaoEncoreBundle")) {
+            $loader->load('@HeimrichHannotContaoMultiFileUploadBundle/Resources/config/config_encore.yml');
+        }
     }
 }